requirements.txt
tkintermapview
matplotlib
pyserial
cobs